requirements.txt
asgiref==3.5.2
astroid==2.5.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-admin-interface==0.19.1
django-colorfield==0.7.1
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.1.0
pylint==2.7.2
pytz==2021.1
PyYAML==6.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
urllib3==1.26.4
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
